Refactor system_info routes into separate file
 - PRO-1411
