+++ device-registry/src/test/scala/org/genivi/sota/device_registry/ResourceSpec.scala
@@ -5,14 +5,17 @@
 package org.genivi.sota.device_registry
 
 import akka.http.scaladsl.server.{Directives, Route}
-import akka.http.scaladsl.testkit.RouteTestTimeout
-import akka.http.scaladsl.testkit.ScalatestRouteTest
+import akka.http.scaladsl.testkit.{RouteTestTimeout, ScalatestRouteTest}
 import cats.data.Xor
 import org.genivi.sota.core.DatabaseSpec
 import org.genivi.sota.data._
+import org.genivi.sota.device_registry.db.DeviceRepository
+import org.genivi.sota.http.UuidDirectives.{allowExtractor, extractUuid}
 import org.genivi.sota.messaging.MessageBus
 import org.scalatest.prop.PropertyChecks
 import org.scalatest.{BeforeAndAfterAll, Matchers, PropSpec, Suite}
+
+import scala.concurrent.Future
 import scala.concurrent.duration._
 
 
@@ -39,6 +42,12 @@ trait ResourceSpec extends
 
   lazy val namespaceExtractor = Directives.provide(defaultNs)
 
+  private val namespaceAuthorizer = allowExtractor(namespaceExtractor, extractUuid, deviceAllowed)
+
+  private def deviceAllowed(deviceId: Uuid): Future[Namespace] = {
+    db.run(DeviceRepository.deviceNamespace(deviceId))
+  }
+
   lazy val messageBus =
     MessageBus.publisher(system, system.settings.config) match {
       case Xor.Right(v) => v
@@ -47,7 +56,7 @@ trait ResourceSpec extends
 
   // Route
   lazy implicit val route: Route =
-    new Routing(namespaceExtractor, messageBus).route
+    new DeviceRegistryRoutes(namespaceExtractor, namespaceAuthorizer, messageBus).route
 }
 
 trait ResourcePropSpec extends PropSpec with ResourceSpec with PropertyChecks
